:recycle: [#4246] Rewrite authentication backend for digid-eh oidc plugin
The upstream library callback view now always sends the request to
an authentication backend, and the authentication backend itself must
decide whether to create a django user, or return a dummy user instead
(or plain reject the authentication flow if relevant).

The plugin backend is now added to the top of the backends to try since
the is-candidate check is inexpensive and highly situational to our
own proxy-models, allowing other OIDC backends that create (staff)
users to 'fall-through' as usual.

The backend ensures that no real django user is created and is set up
to extract all the necessary claims from the OIDC flow to be able to
'authenticate' the user and store the relevant information in the
session, so that (in a future PR) we can record this information in
our authentication info model(s).

It looks up configuration aspects from the respective config model
used in the authentication flow (the mechanism is provided by
mozilla-django-oidc-db), or if no such property exists on the model,
in the Django settings.