:recycle: [#4246] Update url config for OIDC
This adds the single-callback-url configuration for the OIDC-based
plugins and deprecates the plugin-specific callback endpoints.

Ideally, we'll have a single callback endpoint for the admin OIDC,
org-oidc and digid-eherkenning OIDC, but that will have to be done in
a separate PR, ideally in combination with the feature flag mechanism
so that new installations get the new behaviour while existing instances
continue using the legacy behaviour by default to prevent breaking
changes.