:pencil: [#4246] Update changelog and documentation
Documented the default legacy behaviour of OIDC callback endpoints, and
update the configuration docs to make people favour the new settings.

A future PR will use a feature flag approach combined with detection
of new/existing instances to apply the most sensible configuration, so
that we don't need people to opt-in to the envvars.