:recycle: Clean up InvalidPluginConfiguration interface
Added explicit message parameter that can be accessed, instead of
having to rely on exc.args[0] which *could* be empty. Now you must
instantiate this exception with a message argument.

The message is still forwarded to the parent exception in case other
usages still rely on exc.args[0], but this will be refactored later.