:building_construction: [#4246] Adapt auth backend and plugin(s) for authenticaton context
Using a development branch of django-digid-eherkenning, additional
tools are available for the OIDC claim processing which stores the
configuration in that library.

The plugin now better understands which data to expect and is not
responsible for the actual claim extraction/processing, instead we
just perform the translation step to be able to store it in the
database and reconstruct a valid authentication context.