:triangular_flag_on_post: [#4246] Add feature flag to allow lax mode
This should make it possible to use this with existing configurations,
by default we do not apply strict mode. Users can add the feature flag
in the admin too to enable strict mode and test things out without
needing to re-deploy.

Machtigingen require additional claims that are pretty much guaranteed
to be absent with certain identity providers. We will need to evaluate
if we need to provide fields for defaults, or if the OIDC providers
will send the necessary data.