:white_check_mark: [#4246] Add test for missing soft-required claim in strict mode
When strict mode is enabled, all claims marked as required are
actually required and absent claims should lead to authentication
failures.

This responsibility lies with OpenID Connect providers/brokers and/or
correct configuration. It is essential for us to be able to build
authentication context data that satisfies the specified schema.