:lock: [#4246] Ensure that additional identifying attributes are hashed
With the extra authentication context metadata, we are storing
additional identifying attributes that can/should be hashed once a
submissions is completed and registered.

Due to the check constraints, we can not create hashes for empty
values, as that implies a certain authentication mode was used, so
we skip over those fields if their values are empty.