:alien: [#4486] Update eHerkenning integration tests with Signicat
This is partially a re-record of the existing vcr cassettes for tests
that are still relevant, with the new, valid certificates and metadata.

On the other hand, the tests have been deleted that would specify the
eHerkenning LOA on a per-authentication-request basis - this feature
was removed from digid-eherkenning (see PR
https://github.com/maykinmedia/django-digid-eherkenning/pull/65\), as
Signicat reported this is not supported.

This is very confusing, as their simulator/test-environment DOES
support it, however, on production this is not supported.

As a consequence, the tests that would provide a LOA option/override
on a per-form level are no longer representative, nor are the test
cases where LoA requirements change for a form (be it resuming, or
right in the middle of an authentication flow).

The mechanism how this *should* work, is that the LoA is defined on
the service is the service catalogue, and that is what the brokers
effectively check during login. This also implies that we should not
actually perform additional validation if the expected LoA was met,
but that code is still present. As it stands, the eHerkenningConfig
LoA attribute/field is only used for the service catalogue entry
generation. Changing this value (especially to a higher level) without
updating the service catalog will break forms, as the code currently
uses this value as an extra check.