:bug: [#4505] Fix crash for status endpoint rendering with completed submission
Completed submissions (by definition) have their authentication
attributes hashed, so hard asserts on this expectation are too
optimistic. The static variables are retrieved/rendered to template
our the confirmation template.

Using the authentication context data with hashed values is not going
to yield the expected results, but during normal registration with a
backend this will be fine since the hashing only happens after
successful registration. A debug logging statement was added to make
this a bit more clear.