:shrug: [#4051] Skip monaco-json-editor interaction on webkit
I could not for the life of me get this to work, after spending 4 hours
on it trying various permutations of workarounds:

* using a textarea or div that is contenteditable to grab the code from,
  via page.evaluate. Those actually all worked, but the problem wasn't
  getting the code/JSON on the clipboard - navigator.clipboard
  actually works just fine
* trying various combinations to get the code in the editor to clear:
  CTRL+A, CTRL+End (to end of document) followed by CTRL+Shift+Home to
  select everything from end to beginning of document, followed by
  Delete key to delete it -> can't get this selection to apply,
  with or without delays provided in keyboard.press

I've added numerous screenshots and compared them to what Chromium
browsers do, and that is different - the select-all-and-then-delete
to get rid of the existing content just doesn't seem to work. I don't
know if the editor is not properly initialized or events are not
properly dispatched, but I've given up at this point.

So, warning other users via an exception that this won't work on webkit
to save them the same debugging hassle and adding a skip helper for
these tests should still give us the coverage on chromium and firefox,
and manual testing on webkit will have to make do.