:art: [#4492] Remove storing upload IDs in the session
This is no longer necessary now that we relate the temporary upload
to a submission (session) - we can use the submission authorization
checks for upload ownership checks.

This removes a source of race conditions, as we have one less array of
UUIDs to keep track of in the session, and especially file uploads were
susceptible to this when multiple uploads are performed in parallel.