:lipstick: [#4456] Display the environment info in the branding
The header is defined as sticky, so we don't have to mess with an
extra sticky element somewhere on the page, which is tougher than it
seems because some of the form submit rows (submit buttons) are
already sticky to the bottom of the page, and there is overlap with
the Django Debug Toolbar handle in development.

Additionally, on mobile, there already is limited screen space, so we
should be careful.

This approach wraps properly on mobile and narrow viewports and doesn't
take additional space (the header on mobile is quite bad by itself, but
that's out of scope).