:test_tube: [#4502] Add regression test for persisting registration backend key too soon
If the key is persisted during suspension, then this can cause
hard-to-debug situations where the form designer expects the
default (first backend) to be selected because the other
logic rules don't match. However, if it is persisted already,
while there's no logic rule match the wrong backend would be
used because it was the 'right' backend before the changes
in the form data are made after resuming the submission.

The simplest way to protect against this is by disallowing
the field to be set when the submissions is not completed
yet.