:adhesive_bandage: [#4502] Prevent the registration backend being saved too prematurely
Detect if we're about to save the value for the
'finalised_registration_backend_key' field with a
submission that is not completed, and prevent it
by raising an error. This is part of the fix that can
be backported.

Another commit will add a data migration and database
constraint to prevent this from happening.