:sparkles: [#4555] Implement the prefill preview/edit form
When merging the prefill columns, we will need to be able to display the
relevant information. The preview component takes care of this.

If a callback handler is provided, this opts-in to editable mode, and a
handle to edit the prefill configuration is displayed and enabled. On
clicking the edit icon, a modal opens, displaying the fields that used to
be in their own columns/inlined in the variables table.

Any validation errors are passed along and displayed, using the existing
admin integration components. Note that we opted not to use Formik's
validation support yet for a couple of reasons:

* we get the validation errors from the backend, so we'd have to pass them
  to initialErrors to get them to be tracked in the Formik state, but we
  don't run any client-side validation (yet). This offers us no benefits
* the data model of validation errors is not (yet) compatible - formik
  expects a string containing a single error message, while Django/DRF can
  output multiple errors in the form of a list of strings. I consider this
  a shortcoming in Formik/Zod etc.
* we have the support for displaying validation errors in existing
  react components, but not yet for Formik, so it's a lot of work that
  doesn't provide any significant benefits yet.