:children_crossing: [#4555] Merge three prefill columns in variables table into one
This declutters the UI and improves the user experience, as the chances of
a horizontal scrollbar get smaller (but more work is needed in different
aspects).

The prefill information is now displayed in a more natural way in the
variables table, and when possible (user defined vars), they can be edited
in a modal that helps focus.

The dispatch action/event was updated so that we can use the Formik submit
event to treat the prefill configuration (plugin, attribute, identifierRole)
as one unit, instead of dispatching separate change events that would be
persisted even without clicking the save button in the modal. Now, if
the key is the empty string, it means "merge" the value into the target
object.