:sparkles: [#4567] Implement objects API registration option fields as Formik fields
Implemented the fields assuming they are running in a Formik context.

Each field is its own component:

* name, label and help text are baked in
* fields can be re-used between legacy/v2 config
* fields synchronize some state to ensure dropdown UI state matches the
  actual form values state

v2 of the configuration requires some data to be reset/cleared if the
api group and/or the objecttype are changed. For this, an optional
callback is provided as prop through wich the confirmation popup can
be wired up. When this is succesful, the change is applied in the
formik state, otherwise it is not. If no callback is provided, the
change is immediately applied. This mechanism allows us to protect the
form designer against accidental changes that would result in a
frustrating form state to undo.

The labels/language have been updated a bit too be shorter and more
informative, compared to the version in master.