:lipstick: [#4567] Ensure form fields inside a modal have consistent styling
The styling was not picked up before because the modal (by default) is
mounted/appended to the document body, which falls outside of the
django admin change form and thus the styles were not properly
picked up.

At some point, we should probably make the modal a part of another
element, but likely it's better to just wait for the complete new
design implementation instead of spending too much time and energy in
this, so a bandaid fix it is for now.