:sparkles: [#4567] Mimick Django's collapsible fieldset
Django admin supports the collapse/expand functionality on admin
fieldsets, which helps focusing the important fields while still
exposing the finer details without being too overwhelming. This
functionality is now also part of our React component.

This ties in with the validation errors - the fieldset should be
expanded if there are validation errors inside to make them visible.
The user of the fieldset needs to provide the list of field names to
watch for, and the parent component needs to provide the validation
error context so that presence of errors can be detected.

If validation errors relevant to the fieldset are detected, the initial
fieldset collapse state is set to 'open'. The user can still toggle
this in the UI as usual.