:children_crossing: [#4567] Replaced RJSF layout with modal/own components
The configuration options for objects API are no longer displayed
inline, but hidden away behind an additional interaction via a modal.

This allows us to leverage Formik, since we now have an explicit
submit action. It also allows us to treat setting the configuration as
a single transaction before we commit it back to the parent state.

Stories have been updated, moved and added, in particular stories for
validation errors have been added, and stories for the variants of
(known) registration backend plugins.

TODO: add a marker on the button to indicate there are validation
errors.