:card_file_box: [#4577] Add database fields to store catalogue reference
We're making this behaviour opt-in and leave the current/old behaviour
in place so that people have time to update their configurations.

The first step is to be able to point to a particular catalogue, so that
in the next step we can point to a document type within that
catalogue. Validation will be added in the next commits, since it is
layered to get to the end-situation of specifying catalogue and
document type without tying it to a particular version or even instance
of the Catalogi API.

A wizard will be added to facilitate migrating the configuration.