:safety_vest: [#4577] Validate that document types (by description ref) exist
Since we will need to resolve the document type URLs dynamically against
the configured catalogue, the minimum validation we can do is check that
at configure-time, the document types exist.

It's still possible they will not be present in the future due to a
number of reasons (lack of micro service consistency guarantees...),
but we expect the problem detection and reporting via the email
digest to mitigate/monitor this.