:white_check_mark: [#4577] Add test cases for options override behaviour
This is quite complex to identify the proper layers. Within each
configuration layer (model fields vs. serializer options), if
both catalogue and document types are specified, they are
validated within the same layer.

However, if some configuration is specified on the model and not
on the serializer, only the catalogue is 'hoisted', since
technically the document types depend on the catalogue. This means
that if the catalogue is overridden, but a document type is not
specified on the serializer while it is specified on the model,
the default on the model must be ignored (since validation of
that value against the serializer catalogue is... weird).