:sparkles: [#4577] Filter document type versions based on version date
This required adding a mechanism to determine which particular version
of the standard an API implements, since using query params that were
added in newer versions on an older version results in HTTP 400
responses because unknown query parameters are used.

There's quite a bit of checking to do and server providers make
mistakes, so we can now explicitly throw errors that have to do with
the API standard (when we think as a client that the server is
implementing it wrong).

For APIs that don't have this parameter, we do the filtering in Python.