:sparkles: [#4577] Implement resolving the document type
We use the submission completion timestamp to resolve which version
of the document type we need, since submissions may be retried a
couple of days/weeks later than when they were actually submitted. The
idea is that the configuration that was in place when the user
submitted the form must be honored.

It's also possible that runtime errors happen, because a version of the
document type existed when the configuration was done, but no new/
future version was set up anymore after time has passed beyond the
'valid until' date of the last version. We can only bail out then
and report the problem.