:sparkles: [#4577] Implement the dropdowns for document type selection
Retrieving the available catalogues is lifted up into the parent
component, since we need access to the API resource URL of the
selected option, which is not guaranteed purely in the form data (this
will not be the case when existing configuration from the DB is
loaded).

Now the available catalogues is passed as a prop to the catalogue
select instead, and the catalogue query URL is used to retrieve teh
available document types.

The available document types are only retrieved once, and passed as
a prop to the document type dropdown widget, since we resolve all
document types within a single selected catalogue.

Whenever a different catalogue is selected, we reset the document types
that were selected, since there is no guarantee they will exist in the
newly selected catalogue.