:fire: Delete snake_case detection
The save hook would fire a celery task on every save
of a form definition instance - including during test runs (in CI).
This creates a lot of tasks in CI to run by the celery worker, which
we also spin up in different tests to check the health check. It is
revealed in the previous commit that real celery events are being
put in the task queue, which is *not* what you want to be happening
in a test suite, as test isolation is required for reproducible
results/outcomes.

The value of this check had also become very low - we structurally
exclude the 'configuration' key from snake/camel case conversion in
the parser/renderer configuration, and formio itself is also not
consistent with snake case and camel case in its own definitions.

It's also only being emitted to Sentry, and probably nobody is actually
looking into these problems - no bug reports have reached us about it
anyway.