:children_crossing: [#4577] Provide feedback for missing options
In the event that a draft document type was selected and afterwards the
feature flag is disabled, this will cause runtime errors but also lead
to confusing UI when no option is visibly selected, yet a validation
error is displayed. Now a warning is displayed in those situations,
with appropriate context depending on the state of the feature flag
for drafts.

The options in the dropdown now also properly display if document types
are in draft state, to help users make a more informed decision when
the feature flag is enabled.