:children_crossing: [#4606] Move ZGW registration options into modal
Moving the ZGW options into a modal allows us to make use of Formik,
since the fields themselves are now properly isolated and we can
add an explicit 'submit' button to commit the configuration back
to the main state. This applies a similar set up like in the objects
API registration options.

The configuration is now broken into parts too - base ZGW options
go in the first tab, while the case property configuration is
moved to its own tab rather than sticking it in a modal itself (
modal in modal is not a user friendly UI).

I'm not sure if putting the objects API options in a fieldset or
maybe its own tab is the best approach, we can discuss.