:art: De-hypothesize regression test
The original reason failures happened was because the form name was
used to construct the file name of the PDF, which resulted in nested
paths with forward slashes.

The ambition with hypothesis was nice - whatever form name you specify
should not result in broken file saving, but hypothesis comes up with
some very exotic unicode characters that font libraries and/or
weasyprint sometimes cannot handle. These edge cases are not worth
fixing or mitigating with fancy input validating restricting the
unicode range of valid characters, but we also do not want these
flaky test failures.

The test is rewritten without hypothesis to illustrate the original
failure, preventing similar further regressions. If the edge cases
do happen for real, we'll have a stakeholder funding the extra
mitigations needed.