:sparkles: [#4606] Update plugin to dynamically resolve case type
Using the case type identification and submission completion
timestamp, resolve the URL of the correct zaaktype version for the
case creation.

This also means that the 'zaaktype' key cannot be used directly
anymore, it must be taken from the created case (easiest). This was
happening for case properties and case roltype medewerker before.