:green_heart: [#4606] Update existing tests to pass again
The additional config option added is required, which causes tests
to crash with a KeyError.

This will/cannot happen in production because the serializer is
responsible for converting/validating the (incomplete) options, and the
default value (empty string) ensures that the key is present in the
cleaned/validated serializer data.