:bug: [#4680] Handle edge case of parsing empty string as json
This cannot be parsed - empty strings means the reference value is
expected to be empty. For the number components, this means null or
undefined, so we can set it to None in python.

This doesn't appear to break the form builder, in fact, when clearing
the numeric reference value entirely, the entire 'eq' key from the
conditional is removed from the config. This is equivalent to setting
it to None in python, since .get on a missing key returns None.