:children_crossing: [#3705] Stop displaying timestamps in SubmissionStep.__str__
When displaying timestamps, we need to make sure to convert to them
local time and then localize the values for display. However, we can
also just decide *not* to display these timestamps since they don't
add much value in the string representation of a submission step.

Instead, we can include the created_on/modified timestamps directly
in the admin inline, and while we're at it, we can declutter the
admin interface by changing to TabularInline instead of StackedInline.

The _data JSONField is no longer displayed - it has since long been
deprecated and replaced with the variables and will be removed
entirely. Since this JSONField is no longer needed, we can fully
benefit from the tabular inlines instead.