:card_file_box: [#4764] Add database field to store reference to price variable
A third pricing mode for forms is being added, and for this one, we
need to configure on the form which variable may contain the price.

If a variable key is set, this implies that price-form-variable is
in effect. Otherwise, logic rules will be checked, and finally,
the configured product price will be used.

The new property needs to be exposed only in the admin-permissions
flavour of the form detail endpoint.