:art: [#4396] Simplify code execution branches
Seems to pass the tests and simplifies the code a little. The try-except
should not happen because of the way form variables are created/synced
from formio definitions, and if it happens, we want to be notified of
it via error monitoring.

Additionally, we can apply the post-processing special treatment for
component sources, and leave the default behaviour for everything else,
rather than potentially missing some cases if additional sources get
added in the future.

The variable itself is already mutated during the saving of prefill
data, so just updating the dict holding the [key, value] pairs is
sufficient.