:bug: Fix type error in object spread
This caused the preset nested validate schema for components not to be
copied over properly, leading to some accessibility issues, but
probably a whole other set of problems.

I suspect an editor auto-complete thing messing things up, combined
with the object spread of undefined leading to an empty object rather
than crashing hard lead to this bug not being detected/reported in two
years.