:sparkles: [#4320] Build up smarter URLs for cosign start
If URLs in emails are enabled for the cosign request email, create a
more useful URL to the form by giving a 'init-cosign' instruction,
passing along the submission reference/code so that we can auto-populate
that information and cut out intermediate screens for the end-user.

If URLs are not enabled, we stick to the old behaviour.