:sparkles: [#4320] Always emit cosign login options again, different strategy
Now also emit whether links are enabled or not, and let the frontend code
decide based on this information whether options should be rendered
or not. We need access to the options in certain views on the frontend,
and we cannot pass a query param to opt in/out from the frontend as
the API call to retrieve the form is done *after* the param processing.