:sparkles: [#4320] Add ability to add 'cosign now' button on confirmation page
Added a custom templatetag that can optionally be included in the
confirmation page content/template to dislay a link/button that allows
users to start the cosign process from the confirmation page.

This is the same link that is included in emails, when the form_url
variable is used in there. The appearance is fixed to primary button,
the button text can be controlled from the template.