:technologist: Fix recursion issues in component search strategies
The previous commit removed a source of hypothesis flakiness, but these
strategies can be very useful to simulate weird/unexpected components
that may crash the application or migrations, so It Was Time to fix
the recursion issues. Turns out that defining a deferred strategy
was all that was needed to prevent the recursion in hypothesis.

Additionally, some type checker workarounds could be removed from
existing tests :tada: