:bug: [#4320] Ensure that the summary tag is not emitted if there's nothing to report
The confirmation email is only supposed to emit the summary heading
if there's actual data being summarized. This relies on the renderer
having children or not, since the actual table is produced by the
renderer/tree visitor walking over the submission (steps) and associated
form components in the form definition.

The check whether the render has children was faulty since it returned
children that didn't have any grand-children itself, wrongly creating
the impression that there's going to be content output.

There was some old, commented out, prototype code that hinted at this
not being the intent, that was replaced with a proper check.

This also fixes #4831, which is a duplicate of a bullet in the main
cosign improvements ticket.