:boom: [#4320] Remove header from cosign_information tag
The content of the header needs to change depending on the cosign state,
and we now provide more control in the parent templates so the most
flexible solution is to just remove the header from the hardcoded
template.

While we're working on 3.0 which has breaking changes, this is the
best coincidence we can hope for to make this breaking change.