:recycle: [#4431] Refactor the Objects API v2 handler
The code was already starting to look like a rather fine Italian pasta
before the addressNL mapping because of component-specific handling of
values that was necessary, but now with the additional configuration
options specific to the component type, readability and maintainability
was in serious danger.

I've opted to look at the broader structure on what information we
need, what information we have and what it is exactly that we need
to do with the information:

* We know all the variable values (keys and value), coming from user
  input, derived from user input or entirely independent from user
  input.
* When we know the key of a variable and we know the variable source is
  a form component, we can look up the definition of the form
  component and pass it along as additional context
* We know some user input (file uploads) is processed differently _in
  the context of the objects API_ and we have data storage that allows
  us to translate
* We need to normalize some values to make them suitable for use in the
  Objects API, because a JSON Schema contract is used. Looking at you,
  file uploads with multiple=false.
* We may have additional mapping configuration for a specific variable,
  typically because the variable points to a particular component type
  with special requirements (looking at you, addressNL)
* We must be able to construct the record data for the objects API. The
  keys to be used are specified in the mapped variables configuration,
  namely the target paths.
* We must process all the mapped variables to construct the entire
  object to send to the API.
* We can map all kinds of variables (form, user defined, static,
  registration).

So, working backwards, what we needed was for each mapped variable to
make it return instructions which paths of the final object need to
be set to which value. This is solved with AssignmentSpec data classes,
they simply encode the {key: value} information in a well-typed manner.

Next, a single mapping may result in multiple assignments being done
to different parts of the object - this is the newest requirement from
AddressNL type. We can express that as a sequence of AssignmentSpecs.

Finally, we collect all this information, bundle it up and hand it over
to glom which constructs our actual data object, from the mappings.

That leaves us with the requirement of a thing that produces
AssignmentSpec instructions given a mapping configuration and a value.
Because we already know that the value must be transformed depending
on its meaning (=which component produced it), we also pass the
resolved component configuration along for additional context.

For good measure and better code organization, the bulk of the
implementation is moved into handlers/v2.py. More code will be moved
around if this passes the initial reviews.